requirements.txt
ipykernel==6.29.5
ipython==8.29.0
matplotlib==3.9.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.3
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
polars==1.13.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
torch==2.5.1
dvc==3.58.0
dvc-s3==3.2.0
lightning==2.4.0
torchmetrics==1.6.0
wandb==0.19.0
hydra-core==1.3.2